Ytterligere forbedringer rundt adresse
