+++ melosys-eessi-app/src/main/java/no/nav/melosys/eessi/integration/pdl/dto/sed/DnummerRekvisisjonNorskPostboksadresse.java
@@ -1,12 +1,14 @@
 package no.nav.melosys.eessi.integration.pdl.dto.sed;
 
 import lombok.*;
+import no.nav.melosys.eessi.integration.pdl.dto.sed.adresse.Postboksadresse;
 
 @Data
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class DnummerRekvisisjonNorskPostboksadresse extends DnummerRekvisisjonAdresse.Postboksadresse {
+@EqualsAndHashCode(callSuper = true)
+public class DnummerRekvisisjonNorskPostboksadresse extends Postboksadresse {
 
     private String postboks;
     private String postnummer;
