+++ melosys-eessi-app/src/main/java/no/nav/melosys/eessi/integration/pdl/dto/sed/DnummerRekvisisjonNorskVegadresse.java
@@ -1,12 +1,16 @@
 package no.nav.melosys.eessi.integration.pdl.dto.sed;
 
-import lombok.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.EqualsAndHashCode;
+import no.nav.melosys.eessi.integration.pdl.dto.sed.adresse.Vegadresse;
 
 @Data
 @Builder
 @AllArgsConstructor
 @EqualsAndHashCode(callSuper = true)
-public class DnummerRekvisisjonNorskVegadresse extends DnummerRekvisisjonAdresse.Vegadresse {
+public class DnummerRekvisisjonNorskVegadresse extends Vegadresse {
 
     private String adressenavn;
     private String husnummer;
@@ -15,7 +19,6 @@ public class DnummerRekvisisjonNorskVegadresse extends DnummerRekvisisjonAdresse
     private String bruksenhetsnummer;
     private String tilleggsnavn;
     private String matrikkelId;
-
 }
