+++ melosys-eessi-app/src/main/java/no/nav/melosys/eessi/integration/pdl/dto/sed/DnummerRekvisisjonUtenlandskPostboksadresse.java
@@ -1,13 +1,14 @@
 package no.nav.melosys.eessi.integration.pdl.dto.sed;
 
 import lombok.*;
+import no.nav.melosys.eessi.integration.pdl.dto.sed.adresse.Postboksadresse;
 
 @Data
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
 @EqualsAndHashCode(callSuper = true)
-public class DnummerRekvisisjonUtenlandskPostboksadresse extends DnummerRekvisisjonAdresse.Postboksadresse {
+public class DnummerRekvisisjonUtenlandskPostboksadresse extends Postboksadresse {
 
     private String postboksNummerNavn;
     private String regionDistriktOmraade;
