+++ melosys-eessi-app/src/main/java/no/nav/melosys/eessi/integration/pdl/dto/sed/DnummerRekvisisjonUtenlandskVegadresse.java
@@ -1,13 +1,15 @@
 package no.nav.melosys.eessi.integration.pdl.dto.sed;
 
 import lombok.*;
+import no.nav.melosys.eessi.integration.pdl.dto.sed.adresse.DnummerRekvisisjonAdresse;
+import no.nav.melosys.eessi.integration.pdl.dto.sed.adresse.Vegadresse;
 
 @Data
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
 @EqualsAndHashCode(callSuper = true)
-public class DnummerRekvisisjonUtenlandskVegadresse extends DnummerRekvisisjonAdresse.Vegadresse {
+public class DnummerRekvisisjonUtenlandskVegadresse extends Vegadresse {
 
     private String adressenavnNummer;
     private String bygningEtasjeLeilighet;
