+++ melosys-eessi-app/src/main/java/no/nav/melosys/eessi/integration/pdl/dto/sed/adresse/Matrikkelnummer.java
@@ -1,0 +1,13 @@
+package no.nav.melosys.eessi.integration.pdl.dto.sed.adresse;
+
+import lombok.Builder;
+import lombok.Value;
+
+@Value
+@Builder
+public class Matrikkelnummer {
+    String kommunenummer;
+    int gaardsnummer;
+    int bruksnummer;
+    Integer festenummer;
+}
