+++ melosys-eessi-app/src/main/java/no/nav/melosys/eessi/integration/pdl/dto/sed/adresse/Postboksadresse.java
@@ -1,0 +1,12 @@
+package no.nav.melosys.eessi.integration.pdl.dto.sed.adresse;
+
+import lombok.*;
+
+@RequiredArgsConstructor
+@AllArgsConstructor
+public class Postboksadresse {
+
+    String postbokseier;
+    Poststed poststed;
+    String postboks;
+}
