+++ melosys-eessi-app/src/main/java/no/nav/melosys/eessi/integration/pdl/dto/sed/adresse/Poststed.java
@@ -1,0 +1,11 @@
+package no.nav.melosys.eessi.integration.pdl.dto.sed.adresse;
+
+import lombok.Builder;
+import lombok.Value;
+
+@Value
+@Builder
+public class Poststed {
+    String poststedsnavn;
+    String postnummer;
+}
